Refactoring structure.
