+++ dces-drc-integration/src/integrationTest/java/uk/gov/justice/laa/crime/dces/integration/service/spy/SpyFactory.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.integration.testing;
+package uk.gov.justice.laa.crime.dces.integration.service.spy;
 
 import org.springframework.boot.test.mock.mockito.SpyBean;
 import org.springframework.stereotype.Component;
